Refactor load_recordings_from_mbids
