+++ listenbrainz/labs_api/labs/api/recording_from_recording_mbid.py
@@ -2,9 +2,7 @@ import psycopg2.extras
 from datasethoster import Query
 from flask import current_app
 
-from listenbrainz.labs_api.labs.api import utils
-
-psycopg2.extras.register_uuid()
+from listenbrainz.db.recording import load_recordings_from_mbids_with_redirects
 
 
 class RecordingFromRecordingMBIDQuery(Query):
@@ -32,7 +30,7 @@ class RecordingFromRecordingMBIDQuery(Query):
                 psycopg2.connect(current_app.config["SQLALCHEMY_TIMESCALE_URI"]) as ts_conn, \
                 mb_conn.cursor(cursor_factory=psycopg2.extras.DictCursor) as mb_curs, \
                 ts_conn.cursor(cursor_factory=psycopg2.extras.DictCursor) as ts_curs:
-            output = utils.get_recordings_from_mbids(mb_curs, ts_curs, mbids)
+            output = load_recordings_from_mbids_with_redirects(mb_curs, ts_curs, mbids)
 
         # Ideally offset and count should be handled by the postgres query itself, but the 1:1 relationship
         # of what the user requests and what we need to fetch is no longer true, so we can't easily use LIMIT/OFFSET.
